backend/requirements.txt
fastapi
uvicorn
sqlalchemy
PyMuPDF
python-dotenv
langchain-google-genai
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
pydantic[email]
Pillow
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sentence-transformers
langchain-groq
langgraph
supabase
aiofiles
